+++ src/Data/Promotion/Prelude/Function.hs
@@ -28,10 +28,10 @@ module Data.Promotion.Prelude.Function (
     -- * Defunctionalization symbols
   , IdSym0, IdSym1
   , ConstSym0, ConstSym1, ConstSym2
-  , (:.@#$%^%$#@$), (:.@#$%^%$#@$$), (:.@#$%^%$#@$$$), (:.@#$%^%$#@$$$$)
+  , (:.@#@$), (:.@#@$$), (:.@#@$$$), (:.@#@$$$$)
   , FlipSym0, FlipSym1, FlipSym2, FlipSym3
-  , (:$@#$%^%$#@$), (:$@#$%^%$#@$$), (:$@#$%^%$#@$$$)
-  , (:&@#$%^%$#@$), (:&@#$%^%$#@$$), (:&@#$%^%$#@$$$)
+  , (:$@#@$), (:$@#@$$), (:$@#@$$$)
+  , (:&@#@$), (:&@#@$$), (:&@#@$$$)
   , OnSym0, OnSym1, OnSym2, OnSym3, OnSym4
   ) where