+++ src/Data/Singletons/Prelude/Bool.hs
@@ -48,8 +48,8 @@ module Data.Singletons.Prelude.Bool (
   TrueSym0, FalseSym0,
 
   NotSym0, NotSym1,
-  (:&&@#$%^%$#@$), (:&&@#$%^%$#@$$), (:&&@#$%^%$#@$$$),
-  (:||@#$%^%$#@$), (:||@#$%^%$#@$$), (:||@#$%^%$#@$$$),
+  (:&&@#@$), (:&&@#@$$), (:&&@#@$$$),
+  (:||@#@$), (:||@#@$$), (:||@#@$$$),
   Bool_Sym0, Bool_Sym1, Bool_Sym2, Bool_Sym3,
   OtherwiseSym0
   ) where